requirements.txt
paho-mqtt==1.6.1
aiogram==3.15.0
